requirements.txt
flask
flask-jwt-extended
flask-pymongo
flask-cors
bcrypt
python-dotenv
dnspython